ArcGIS Project
The goal of this project is find a site for a new school in central
Ottawa area. The new school location should not close to current
existing school, but it should near transit route and near residential
area. Also, the area in Ottawa
that have higher amounts of population of age under 18 should have
higher  demand for school, so this area will be a better location for
building new  school. The result of my project will produce a raster
file(.tif) that indicate the suitable area for a new school, the range
of this area are all good to build a new school, but if the value of
location is higher, that means it is more appropriate location to build
a new school.